Add a "project" argument to Pytest.vim
This runs tests for the entire project. It does this by starting in the
current directory and walking up to '/' looking for a directory called 'tests'
or a file called 'tests.py'. If it finds either of these, run py.test
otherwise error out letting the user know that there are no tests defined for
the project tree. There is a helper argument 'showprojecttests' which shows
where in the directory tree we will run py.test